Libretro reorganizing
