+++ libretro/libretro-mapper.c
@@ -15,7 +15,7 @@
 #include "kbd.h"
 #include "mousedrv.h"
 
-static retro_input_state_t input_state_cb;
+retro_input_state_t input_state_cb;
 static retro_input_poll_t input_poll_cb;
 
 void retro_set_input_state(retro_input_state_t cb)
@@ -27,21 +27,6 @@ void retro_set_input_poll(retro_input_poll_t cb)
 {
    input_poll_cb = cb;
 }
-
-#ifdef POINTER_DEBUG
-int pointer_x = 0;
-int pointer_y = 0;
-#endif
-int last_pointer_x = 0;
-int last_pointer_y = 0;
-
-/* VKBD starting point: 10x3 == f7 */
-int vkey_pos_x = 10;
-int vkey_pos_y = 3;
-int vkbd_x_min = 0;
-int vkbd_x_max = 0;
-int vkbd_y_min = 0;
-int vkbd_y_max = 0;
 
 /* Mouse speed flags */
 #define MOUSE_SPEED_SLOWER 1
@@ -63,9 +48,10 @@ static void retro_mouse_move(int x, int y)
 
 /* Core flags */
 int mapper_keys[RETRO_MAPPER_LAST] = {0};
-static int mapper_flag[3][128] = {0};
-int vkflag[10] = {0};
 int retro_capslock = false;
+static int mapper_flag[3][128] = {0};
+extern int vkflag[10];
+
 unsigned int cur_port = 2;
 static int cur_port_prev = -1;
 bool cur_port_locked = false;
@@ -78,11 +64,11 @@ unsigned int retro_warpmode = 0;
 extern bool retro_vkbd;
 extern bool retro_vkbd_transparent;
 extern short int retro_vkbd_ready;
-extern unsigned int retro_devices[RETRO_DEVICES];
+extern void input_vkbd(void);
 
 static unsigned retro_key_state[RETROK_LAST] = {0};
 unsigned retro_key_event_state[RETROK_LAST] = {0};
-static int16_t joypad_bits[RETRO_DEVICES];
+int16_t joypad_bits[RETRO_DEVICES];
 extern bool libretro_supports_bitmasks;
 extern dc_storage *dc;
 
@@ -147,43 +133,6 @@ int retro_ui_get_pointer_state(int *px, int *py, unsigned int *pbuttons)
 
    return 1;
 }
-
-enum EMU_FUNCTIONS
-{
-   EMU_VKBD = 0,
-   EMU_STATUSBAR,
-   EMU_JOYPORT,
-   EMU_RESET,
-   EMU_SAVE_DISK,
-   EMU_ASPECT_RATIO,
-   EMU_ZOOM_MODE,
-   EMU_TURBO_FIRE,
-   EMU_WARP_MODE,
-   EMU_DATASETTE_HOTKEYS,
-   EMU_DATASETTE_STOP,
-   EMU_DATASETTE_START,
-   EMU_DATASETTE_FORWARD,
-   EMU_DATASETTE_REWIND,
-   EMU_DATASETTE_RESET,
-   EMU_FUNCTION_COUNT
-};
-
-/* VKBD_MIN_HOLDING_TIME: Hold a direction longer than this and automatic movement sets in */
-/* VKBD_MOVE_DELAY: Delay between automatic movement from button to button */
-#define VKBD_MIN_HOLDING_TIME 200
-#define VKBD_MOVE_DELAY 50
-bool let_go_of_direction = true;
-long last_move_time = 0;
-long last_press_time = 0;
-
-/* VKBD_STICKY_HOLDING_TIME: Button press longer than this triggers sticky key */
-#define VKBD_STICKY_HOLDING_TIME 1000
-int let_go_of_button = 1;
-long last_press_time_button = 0;
-int vkey_pressed = -1;
-int vkey_sticky = -1;
-int vkey_sticky1 = -1;
-int vkey_sticky2 = -1;
 
 void emu_function(int function)
 {
@@ -357,62 +306,21 @@ void update_input(unsigned disable_keys)
     * INDEX  0  1  2  3  4  5  6  7  8  9  10  11  12  13  14  15  16  17  18  19  20  21  22  23
     */
 
-   static long now = 0;
-   static long last_vkey_pressed_time = 0;
-   static int last_vkey_pressed = -1;
-   static int vkey_sticky1_release = 0;
-   static int vkey_sticky2_release = 0;
-
-   static int i = 0, j = 0, mk = 0;
-   static int LX = 0, LY = 0, RX = 0, RY = 0;
+   unsigned int i = 0, j = 0, mk = 0;
+   int LX = 0, LY = 0, RX = 0, RY = 0;
    const int threshold = 20000;
 
-   static int jbt[2][24] = {0};
+   static int jbt[2][RETRO_DEVICE_ID_JOYPAD_LAST] = {0};
    static int kbt[EMU_FUNCTION_COUNT] = {0};
     
-   now = retro_ticks() / 1000;
-
-   if (vkey_sticky && last_vkey_pressed != -1 && last_vkey_pressed > 0)
-   {
-      if (vkey_sticky1 > -1 && vkey_sticky1 != last_vkey_pressed)
-      {
-         if (vkey_sticky2 > -1 && vkey_sticky2 != last_vkey_pressed)
-            kbd_handle_keyup(vkey_sticky2);
-         vkey_sticky2 = last_vkey_pressed;
-      }
-      else
-         vkey_sticky1 = last_vkey_pressed;
-   }
-
-   /* Keyup only after button is up */
-   if (last_vkey_pressed != -1 && !vkflag[RETRO_DEVICE_ID_JOYPAD_B])
-   {
-      if (vkey_pressed == -1 && last_vkey_pressed >= 0 && last_vkey_pressed != vkey_sticky1 && last_vkey_pressed != vkey_sticky2)
-         kbd_handle_keyup(last_vkey_pressed);
-
-      last_vkey_pressed = -1;
-   }
-
-   if (vkey_sticky1_release)
-   {
-      vkey_sticky1_release = 0;
-      vkey_sticky1 = -1;
-      kbd_handle_keyup(vkey_sticky1);
-   }
-   if (vkey_sticky2_release)
-   {
-      vkey_sticky2_release = 0;
-      vkey_sticky2 = -1;
-      kbd_handle_keyup(vkey_sticky2);
-   }
-
    /* Iterate hotkeys, skip Datasette hotkeys if Datasette hotkeys are disabled or if VKBD is on */
    int i_last = (datasette_hotkeys && !retro_vkbd) ? RETRO_MAPPER_DATASETTE_RESET : RETRO_MAPPER_DATASETTE_HOTKEYS;
-   i_last -= 24;
+   i_last -= RETRO_DEVICE_ID_JOYPAD_LAST;
 
    for (i = 0; i <= i_last; i++)
    {
-      mk = i + 24; /* Skip RetroPad mappings from mapper_keys */
+      /* Skip RetroPad mappings from mapper_keys */
+      mk = i + RETRO_DEVICE_ID_JOYPAD_LAST;
 
       /* Key down */
       if (input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, mapper_keys[mk]) && !kbt[i] && mapper_keys[mk])
@@ -707,298 +615,7 @@ void update_input(unsigned disable_keys)
    } /* for j */
 
    /* Virtual keyboard for ports 1 & 2 */
-   if (retro_vkbd)
-   {
-      /* Wait for all inputs to be released */
-      if (retro_vkbd_ready < 1)
-      {
-         if ((retro_vkbd_ready == 0 && !joypad_bits[0] && !joypad_bits[1]) ||
-              retro_vkbd_ready < 0)
-            retro_vkbd_ready++;
-         return;
-      }
-
-      if (!vkflag[RETRO_DEVICE_ID_JOYPAD_B]) /* Allow directions when key is not pressed */
-      {
-         if (!vkflag[RETRO_DEVICE_ID_JOYPAD_UP] && ((joypad_bits[0] & (1 << RETRO_DEVICE_ID_JOYPAD_UP)) ||
-                                                    (joypad_bits[1] & (1 << RETRO_DEVICE_ID_JOYPAD_UP)) ||
-                                                    input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_UP)))
-            vkflag[RETRO_DEVICE_ID_JOYPAD_UP] = 1;
-         else
-         if (vkflag[RETRO_DEVICE_ID_JOYPAD_UP] && (!(joypad_bits[0] & (1 << RETRO_DEVICE_ID_JOYPAD_UP)) &&
-                                                   !(joypad_bits[1] & (1 << RETRO_DEVICE_ID_JOYPAD_UP)) &&
-                                                   !input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_UP)))
-            vkflag[RETRO_DEVICE_ID_JOYPAD_UP] = 0;
-
-         if (!vkflag[RETRO_DEVICE_ID_JOYPAD_DOWN] && ((joypad_bits[0] & (1 << RETRO_DEVICE_ID_JOYPAD_DOWN)) ||
-                                                      (joypad_bits[1] & (1 << RETRO_DEVICE_ID_JOYPAD_DOWN)) ||
-                                                      input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_DOWN)))
-            vkflag[RETRO_DEVICE_ID_JOYPAD_DOWN] = 1;
-         else
-         if (vkflag[RETRO_DEVICE_ID_JOYPAD_DOWN] && (!(joypad_bits[0] & (1 << RETRO_DEVICE_ID_JOYPAD_DOWN)) &&
-                                                     !(joypad_bits[1] & (1 << RETRO_DEVICE_ID_JOYPAD_DOWN)) &&
-                                                     !input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_DOWN)))
-            vkflag[RETRO_DEVICE_ID_JOYPAD_DOWN] = 0;
-
-         if (!vkflag[RETRO_DEVICE_ID_JOYPAD_LEFT] && ((joypad_bits[0] & (1 << RETRO_DEVICE_ID_JOYPAD_LEFT)) ||
-                                                      (joypad_bits[1] & (1 << RETRO_DEVICE_ID_JOYPAD_LEFT)) ||
-                                                      input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_LEFT)))
-            vkflag[RETRO_DEVICE_ID_JOYPAD_LEFT] = 1;
-         else
-         if (vkflag[RETRO_DEVICE_ID_JOYPAD_LEFT] && (!(joypad_bits[0] & (1 << RETRO_DEVICE_ID_JOYPAD_LEFT)) &&
-                                                     !(joypad_bits[1] & (1 << RETRO_DEVICE_ID_JOYPAD_LEFT)) &&
-                                                     !input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_LEFT)))
-            vkflag[RETRO_DEVICE_ID_JOYPAD_LEFT] = 0;
-
-         if (!vkflag[RETRO_DEVICE_ID_JOYPAD_RIGHT] && ((joypad_bits[0] & (1 << RETRO_DEVICE_ID_JOYPAD_RIGHT)) ||
-                                                       (joypad_bits[1] & (1 << RETRO_DEVICE_ID_JOYPAD_RIGHT)) ||
-                                                       input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_RIGHT)))
-            vkflag[RETRO_DEVICE_ID_JOYPAD_RIGHT] = 1;
-         else
-         if (vkflag[RETRO_DEVICE_ID_JOYPAD_RIGHT] && (!(joypad_bits[0] & (1 << RETRO_DEVICE_ID_JOYPAD_RIGHT)) &&
-                                                      !(joypad_bits[1] & (1 << RETRO_DEVICE_ID_JOYPAD_RIGHT)) &&
-                                                      !input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_RIGHT)))
-            vkflag[RETRO_DEVICE_ID_JOYPAD_RIGHT] = 0;
-      }
-      else /* Release all directions when key is pressed */
-      {
-         vkflag[RETRO_DEVICE_ID_JOYPAD_UP]    = 0;
-         vkflag[RETRO_DEVICE_ID_JOYPAD_DOWN]  = 0;
-         vkflag[RETRO_DEVICE_ID_JOYPAD_LEFT]  = 0;
-         vkflag[RETRO_DEVICE_ID_JOYPAD_RIGHT] = 0;
-      }
-
-      if (vkflag[RETRO_DEVICE_ID_JOYPAD_UP] ||
-          vkflag[RETRO_DEVICE_ID_JOYPAD_DOWN] ||
-          vkflag[RETRO_DEVICE_ID_JOYPAD_LEFT] ||
-          vkflag[RETRO_DEVICE_ID_JOYPAD_RIGHT])
-      {
-         if (let_go_of_direction)
-            /* just pressing down */
-            last_press_time = now;
-
-         if ((now - last_press_time > VKBD_MIN_HOLDING_TIME
-           && now - last_move_time > VKBD_MOVE_DELAY)
-           || let_go_of_direction)
-         {
-            last_move_time = now;
-
-            if (vkflag[RETRO_DEVICE_ID_JOYPAD_UP])
-               vkey_pos_y -= 1;
-            else if (vkflag[RETRO_DEVICE_ID_JOYPAD_DOWN])
-               vkey_pos_y += 1;
-
-            if (vkflag[RETRO_DEVICE_ID_JOYPAD_LEFT])
-               vkey_pos_x -= 1;
-            else if (vkflag[RETRO_DEVICE_ID_JOYPAD_RIGHT])
-               vkey_pos_x += 1;
-         }
-         let_go_of_direction = false;
-      }
-      else
-         let_go_of_direction = true;
-
-      if (vkey_pos_x < 0)
-         vkey_pos_x = VKBDX - 1;
-      else if (vkey_pos_x > VKBDX - 1)
-         vkey_pos_x = 0;
-      if (vkey_pos_y < 0)
-         vkey_pos_y = VKBDY - 1;
-      else if (vkey_pos_y > VKBDY - 1)
-         vkey_pos_y = 0;
-
-      /* Absolute pointer */
-      int p_x = input_state_cb(0, RETRO_DEVICE_POINTER, 0, RETRO_DEVICE_ID_POINTER_X);
-      int p_y = input_state_cb(0, RETRO_DEVICE_POINTER, 0, RETRO_DEVICE_ID_POINTER_Y);
-
-      if (p_x != 0 && p_y != 0 && (p_x != last_pointer_x || p_y != last_pointer_y))
-      {
-         int px = (int)((p_x + 0x7fff) * retrow / 0xffff);
-         int py = (int)((p_y + 0x7fff) * retroh / 0xffff);
-         last_pointer_x = p_x;
-         last_pointer_y = p_y;
-#ifdef POINTER_DEBUG
-         pointer_x = px;
-         pointer_y = py;
-#endif
-         if (px >= vkbd_x_min && px <= vkbd_x_max && py >= vkbd_y_min && py <= vkbd_y_max)
-         {
-            float vkey_width = (float)(vkbd_x_max - vkbd_x_min) / VKBDX;
-            vkey_pos_x = ((px - vkbd_x_min) / vkey_width);
-
-            float vkey_height = (float)(vkbd_y_max - vkbd_y_min) / VKBDY;
-            vkey_pos_y = ((py - vkbd_y_min) / vkey_height);
-
-            vkey_pos_x = (vkey_pos_x < 0) ? 0 : vkey_pos_x;
-            vkey_pos_x = (vkey_pos_x > VKBDX - 1) ? VKBDX - 1 : vkey_pos_x;
-            vkey_pos_y = (vkey_pos_y < 0) ? 0 : vkey_pos_y;
-            vkey_pos_y = (vkey_pos_y > VKBDY - 1) ? VKBDY - 1 : vkey_pos_y;
-
-#ifdef POINTER_DEBUG
-            printf("px:%d py:%d (%d,%d) vkey:%dx%d\n", p_x, p_y, px, py, vkey_pos_x, vkey_pos_y);
-#endif
-         }
-      }
-
-      /* Press Return, RetroPad Start */
-      i = RETRO_DEVICE_ID_JOYPAD_START;
-      if (!vkflag[i] && mapper_keys[i] >= 0 && ((joypad_bits[0] & (1 << i)) ||
-                                                (joypad_bits[1] & (1 << i)))
-                                            && !input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_RETURN))
-      {
-         vkflag[i] = 1;
-         kbd_handle_keydown(RETROK_RETURN);
-      }
-      else
-      if (vkflag[i] && (!(joypad_bits[0] & (1 << i)) &&
-                        !(joypad_bits[1] & (1 << i))))
-      {
-         vkflag[i] = 0;
-         kbd_handle_keyup(RETROK_RETURN);
-      }
-
-      /* Toggle ShiftLock, RetroPad Y */
-      i = RETRO_DEVICE_ID_JOYPAD_Y;
-      if (!vkflag[i] && mapper_keys[i] >= 0 && ((joypad_bits[0] & (1 << i)) ||
-                                                (joypad_bits[1] & (1 << i))))
-      {
-         vkflag[i] = 1;
-         retro_key_down(RETROK_CAPSLOCK);
-         retro_key_up(RETROK_CAPSLOCK);
-      }
-      else
-      if (vkflag[i] && (!(joypad_bits[0] & (1 << i)) &&
-                        !(joypad_bits[1] & (1 << i))))
-      {
-         vkflag[i] = 0;
-      }
-
-      /* Press Space, RetroPad X */
-      i = RETRO_DEVICE_ID_JOYPAD_X;
-      if (!vkflag[i] && mapper_keys[i] >= 0 && ((joypad_bits[0] & (1 << i)) ||
-                                                (joypad_bits[1] & (1 << i))))
-      {
-         vkflag[i] = 1;
-         kbd_handle_keydown(RETROK_SPACE);
-      }
-      else
-      if (vkflag[i] && (!(joypad_bits[0] & (1 << i)) &&
-                        !(joypad_bits[1] & (1 << i))))
-      {
-         vkflag[i] = 0;
-         kbd_handle_keyup(RETROK_SPACE);
-      }
-
-      /* Toggle transparency, RetroPad A */
-      i = RETRO_DEVICE_ID_JOYPAD_A;
-      if (!vkflag[i] && mapper_keys[i] >= 0 && ((joypad_bits[0] & (1 << i)) ||
-                                                (joypad_bits[1] & (1 << i))))
-      {
-         vkflag[i] = 1;
-         retro_vkbd_transparent = !retro_vkbd_transparent;
-      }
-      else
-      if (vkflag[i] && (!(joypad_bits[0] & (1 << i)) &&
-                        !(joypad_bits[1] & (1 << i))))
-      {
-         vkflag[i] = 0;
-      }
-
-      /* Key press, RetroPad B joyports 1+2 / Keyboard Enter / Pointer */
-      i = RETRO_DEVICE_ID_JOYPAD_B;
-      if (!vkflag[i] && ((joypad_bits[0] & (1 << i)) ||
-                         (joypad_bits[1] & (1 << i)) ||
-                         input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_RETURN) ||
-                         input_state_cb(0, RETRO_DEVICE_POINTER, 0, RETRO_DEVICE_ID_POINTER_PRESSED)))
-      {
-         vkey_pressed = vkeys[(vkey_pos_y * VKBDX) + vkey_pos_x + ((0) ? VKBDX * VKBDY : 0)].value;
-
-         vkflag[i] = 1;
-
-         if (vkey_pressed != -1 && last_vkey_pressed == -1)
-         {
-            switch (vkey_pressed)
-            {
-               case -2:
-                  emu_function(EMU_RESET);
-                  break;
-               case -3:
-                  if (retro_capslock)
-                     emu_function(EMU_SAVE_DISK);
-                  else
-                     emu_function(EMU_STATUSBAR);
-                  break;
-               case -4:
-                  if (retro_capslock)
-                     emu_function(EMU_ASPECT_RATIO);
-                  else
-                     emu_function(EMU_JOYPORT);
-                  break;
-               case -5:
-                  if (retro_capslock)
-                     emu_function(EMU_ZOOM_MODE);
-                  else
-                     emu_function(EMU_TURBO_FIRE);
-                  break;
-               case -10: /* ShiftLock */
-                  retro_key_down(RETROK_CAPSLOCK);
-                  retro_key_up(RETROK_CAPSLOCK);
-                  break;
-
-               case -11:
-                  emu_function(EMU_DATASETTE_STOP);
-                  break;
-               case -12:
-                  emu_function(EMU_DATASETTE_START);
-                  break;
-               case -13:
-                  emu_function(EMU_DATASETTE_FORWARD);
-                  break;
-               case -14:
-                  emu_function(EMU_DATASETTE_REWIND);
-                  break;
-               case -15:
-                  emu_function(EMU_DATASETTE_RESET);
-                  break;
-
-               default:
-                  if (vkey_pressed == vkey_sticky1)
-                     vkey_sticky1_release = 1;
-                  if (vkey_pressed == vkey_sticky2)
-                     vkey_sticky2_release = 1;
-                  kbd_handle_keydown(vkey_pressed);
-                  break;
-            }
-         }
-         last_vkey_pressed = vkey_pressed;
-      }
-      else
-      if (vkflag[i] && (!(joypad_bits[0] & (1 << i)) &&
-                        !(joypad_bits[1] & (1 << i)) &&
-                        !input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_RETURN) &&
-                        !input_state_cb(0, RETRO_DEVICE_POINTER, 0, RETRO_DEVICE_ID_POINTER_PRESSED)))
-      {
-         vkey_pressed = -1;
-         vkflag[i] = 0;
-      }
-
-      if (vkflag[RETRO_DEVICE_ID_JOYPAD_B] && vkey_pressed > 0)
-      {
-         if (let_go_of_button)
-            last_press_time_button = now;
-         if (now - last_press_time_button > VKBD_STICKY_HOLDING_TIME)
-            vkey_sticky = 1;
-         let_go_of_button = 0;
-      }
-      else
-      {
-         let_go_of_button = 1;
-         vkey_sticky = 0;
-      }
-   }
-#if 0
-   printf("vkey:%d sticky:%d sticky1:%d sticky2:%d, now:%d last:%d\n", vkey_pressed, vkey_sticky, vkey_sticky1, vkey_sticky2, now, last_press_time_button);
-#endif
+   input_vkbd();
 }
 
 int process_keyboard_pass_through()
